+++ packages/hedron-django/src/hedron_django/__init__.py
@@ -16,7 +16,7 @@ from hedron_django.responses import component_response, interaction_response
 from hedron_django.routing import DjangoUrlReverser, hedron_view
 from hedron_django.urls import component_path, hedron_paths, include_component_path
 
-__version__ = "0.24.0"
+__version__ = "0.25.0"
 
 __all__ = [
     "DjangoUrlReverser",